requirements.txt
git+https://github.com/Pycord-Development/pycord
imagehash~=4.3.1
pillow~=9.2.0
requests~=2.28.1
wheel
python-dotenv~=0.21.0
numpy~=1.23.4